pyproject.toml
[project]
name = "new-folder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.7.0",
    "google-generativeai>=0.3.0",
    "joblib>=1.0.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "python-dotenv>=0.19.0",
    "pyyaml>=5.4.0",
    "rapidfuzz>=3.0.0",
    "scikit-learn>=0.24.0",
    "scipy>=1.7.0",
    "sentence-transformers>=2.0.0",
    "streamlit>=1.36.0",
    "torch>=2.0.0",
    "transformers>=4.30.0",
]
